requirements.txt
numpy
cupy-cuda12x
pydub
soundfile
mutagen
scipy
logging